+++ src/starkbank/transfer.clj
@@ -9,10 +9,12 @@
     - `:tax-id` [string]: receiver tax ID (CPF or CNPJ) with or without formatting. ex: \"01234567890\" or \"20.018.183/0001-80\"
     - `:bank-code` [string]: code of the receiver bank institution in Brazil. If an ISPB (8 digits) is informed, a PIX transfer will be created, else a TED will be issued. ex: \"20018183\" or \"341\"
     - `:branch-code` [string]: receiver bank account branch. Use '-' in case there is a verifier digit. ex: \"1357-9\"
-    - `:account-number` [string]: Receiver Bank Account number. Use '-' before the verifier digit. ex: \"876543-2\"
+    - `:account-number` [string]: receiver bank account number. Use '-' before the verifier digit. ex: \"876543-2\"
 
   ## Parameters (optional):
-    - `:scheduled` [string]: date or datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: \"2021-03-11T08:00:00.000000+00:00\"
+    - `:account-type` [string, \"checking\"]: receiver bank account type. This parameter only has effect on Pix Transfers. ex: \"checking\", \"savings\" or \"salary\"
+    - `:external-id` [string, default nil]: url safe string that must be unique among all your transfers. Duplicated external_ids will cause failures. By default, this parameter will block any transfer that repeats amount and receiver information on the same date. ex: \"my-internal-id-123456\"
+    - `:scheduled` [string, default now]: date or datetime when the transfer will be processed. May be pushed to next business day if necessary. ex: \"2021-03-11T08:00:00.000000+00:00\"
     - `:tags` [list of strings]: list of strings for reference when searching for transfers. ex: [\"employees\", \"monthly\"]
 
   Attributes (return-only):
@@ -36,6 +38,8 @@
       bank-code "bank-code"
       branch-code "branch-code"
       account-number "account-number"
+      account-type "account-type"
+      external-id "external-id"
       scheduled "scheduled"
       tags "tags"
     }
@@ -49,6 +53,8 @@
           "bankCode" bank-code
           "branchCode" branch-code
           "accountNumber" account-number
+          "accountType" account-type
+          "externalId" external-id
           "scheduled" scheduled
           "tags" (if (nil? tags) nil (into-array String tags))
         }
@@ -64,6 +70,8 @@
       :bank-code (.bankCode java-object)
       :branch-code (.branchCode java-object)
       :account-number (.accountNumber java-object)
+      :account-type (.accountType java-object)
+      :external-id (.externalId java-object)
       :scheduled (.scheduled java-object)
       :tags (into [] (.tags java-object))
       :fee (.fee java-object)